VK_foto/requirements.txt
requests
python-dotenv
time
PySimpleGUI
operator